fix le mapper des UE et Student j'avais oublié une méthode + typage du coeff dans le dto des ue

+++ src/main/java/org/mmi/MMIPlatform/application/dto/UEDto.java
@@ -17,7 +17,7 @@ public class UEDto {
     @JsonProperty("sum_note")
     private double sum_note;
     @JsonProperty("coeff")
-    private double coeff;
+    private float coeff;
 
 
 }
